+++ feature/profile/build.gradle.kts
@@ -10,4 +10,6 @@ android {
 dependencies {
     implementation(projects.core.data)
     implementation(projects.core.domain)
+
+    implementation(projects.feature.postDetail)
 }
