+++ feature/profile/src/main/kotlin/com/site/pochak/app/feature/profile/ProfileScreen.kt
@@ -44,6 +44,10 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavHost
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
 import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
@@ -59,6 +63,7 @@ import com.site.pochak.app.core.designsystem.icon.PochakIcons
 import com.site.pochak.app.core.designsystem.theme.Yellow00
 import com.site.pochak.app.core.model.data.Post
 import com.site.pochak.app.core.network.model.NetworkProfile
+import com.site.pochak.app.feature.profile.navigation.ProfileRoute
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonDecoder
